Add files via upload
I've started finding my "tone" in coding like a musician does as they develop. If you want to use my code, you're going to learn from it, or at least learn what I did along the way. For my animation, the meme image file is Feral and loving it.webp. That was not planned that way intentionally, but that is the image file.  Contrary to how concise the code ended up, this was a learning exercise in the art of "less is more". Overall, it was a good experience (albeit an uncomfortable one) and I hope it can bring value to anyone who happens across it. Enjoy!

-Nate